WIP: Clean up redux actions, thunks, reducer
