+++ .eslintrc
@@ -1,0 +1,7 @@
+{
+  "extends": "react-app",
+  "rules": {
+    "no-unused-vars": "error",
+    "sort-vars": "error"
+  }
+}
