+++ package.json
@@ -16,6 +16,7 @@
     "react-redux": "^5.0.7",
     "react-scripts": "1.1.1",
     "redux": "^3.7.2",
+    "redux-actions": "^2.3.0",
     "redux-devtools-extension": "^2.13.2",
     "redux-storage": "^4.1.2",
     "redux-storage-decorator-filter": "^1.1.8",
