+++ src/redux/index.js
@@ -3,4 +3,5 @@ import configureStore, { getStoredState } from './configureStore';
 export default configureStore;
 export * from './actions';
 export * from './selectors';
+export * from './thunks';
 export { getStoredState };
