--- src/services/util.js
@@ -1,5 +1,0 @@
-import { credentialsSelector, getStoredState } from '../redux';
-
-const readStoredCredentials = () => credentialsSelector(getStoredState());
-
-export { readStoredCredentials };
